+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 2.2.9 (...)
 - expose `default_gateway` in `/profile_list` indicating whether or not the 
   profile expects all traffic over the VPN
+- CSS cache busting in case CSS files change
 
 ## 2.2.8 (2020-05-26)
 - update `PhpSamlSpAuthentication` for php-saml-sp API changes